merging old card service to new one

+++ src/main/java/com/cards/card/repository/CardRepository.java
@@ -3,7 +3,8 @@ package com.cards.card.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.cards.card.entity.CardEntity;
+import java.math.BigInteger;
 
-public interface CardRepository extends MongoRepository<CardEntity, Integer> {
+public interface CardRepository extends MongoRepository<CardEntity, BigInteger> {
 
 }
